refactor: main에만 종속되어있는 컴포넌트 구조 변경
