--- src/components/Journey/Journey.stories.tsx
@@ -1,25 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react';
-
-import { Journey } from './index';
-
-const meta: Meta<typeof Journey> = {
-  title: 'components/Journey',
-  component: Journey,
-  args: {},
-};
-
-export default meta;
-
-type Story = StoryObj<typeof Journey>;
-
-export const Primary: Story = {
-  render: () => (
-    <div
-      css={theme => ({
-        backgroundColor: theme.colors.black800,
-      })}
-    >
-      <Journey />
-    </div>
-  ),
-};
