--- src/components/Journey/index.tsx
@@ -1,1 +1,0 @@
-export { Journey } from './Journey';
