+++ src/components/Main/Journey.tsx
