+++ src/components/Main/RecruitTextSection.tsx
@@ -1,0 +1,11 @@
+export function RecruitTextSection() {
+  return (
+    <section>
+      <h1>
+        서비스 런칭부터 운영까지 <br />
+        열정적인 여정에 지금 합류하세요
+      </h1>
+      <button>14기 지원하기</button>
+    </section>
+  );
+}
