+++ src/components/Main/index.tsx
@@ -1,0 +1,2 @@
+export { RecruitTextSection } from './RecruitTextSection';
+export { SignImageSection } from './SignImageSection';
