+++ src/pages/index.tsx
@@ -1,15 +1,14 @@
-import Image from 'next/image';
 import { css, Theme } from '@emotion/react';
 
 import { FAQ } from '~/components/FAQ';
-import { Journey } from '~/components/Journey';
-import { RecruitEntrance } from '~/components/RecruitEntrance';
+import { RecruitTextSection, SignImageSection } from '~/components/Main';
+import { Journey } from '~/components/Main/Journey';
+import { RecruitEntrance } from '~/components/Main/RecruitEntrance';
 import { ScheduleSection } from '~/components/ScheduleSection';
 import { SectionTitle } from '~/components/SectionTitle';
 import { SEO } from '~/components/SEO';
 import { TimerContainer } from '~/components/TimerContainer';
 import { MEMBER_SCHEDULE, SESSION_SCHEDULES } from '~/constant/schedule';
-import { commonLayoutCss } from '~/styles/layout';
 import { mediaQuery } from '~/styles/media';
 
 export default function Root() {
@@ -29,9 +28,8 @@ export default function Root() {
           <section>
             <FAQ />
           </section>
-          <section css={[commonLayoutCss, signImageContainerCss]}>
-            <Image src="/images/sign/main.png" fill alt="DPM 14th" />
-          </section>
+          <SignImageSection />
+          <RecruitTextSection />
         </div>
       </main>
     </>
@@ -51,22 +49,5 @@ const contentCss = css`
     & > section {
       margin-bottom: 120px;
     }
-  }
-`;
-
-const signImageContainerCss = css`
-  position: relative;
-  width: 100%;
-  height: 356px;
-  img {
-    width: 100%;
-    object-fit: contain;
-  }
-
-  ${mediaQuery('tablet')} {
-    height: 262px;
-  }
-  ${mediaQuery('mobile')} {
-    height: 122px;
   }
 `;
